+++ Demo/Measure-Property.test.ps1
@@ -12,40 +12,69 @@ task UnknownArguments {
 
 task ObjectProperty {
 	# PSObject with Name and Version, both strings
-	$ps = 1 | Select-Object Name, Version
-	$ps.Name = 'string'
-	$ps.Version = 'string'
+	$ps1 = 1 | Select-Object Name, Version
+	$ps1.Name = 'string'
+	$ps1.Version = 'string'
 
-	# Two objects, $host also has Name and Version, Version is not string
-	$r = $Host, $ps | Measure-Property.ps1
-	$r
-	assert ($r[0].Name -eq '' -and $r[0].Type -eq 'InternalHost' -and $r[0].Count -eq 1)
-	assert ($r[1].Name -eq '' -and $r[1].Type -eq 'PSCustomObject' -and $r[1].Count -eq 1)
-	assert ($r[6].Name -eq 'Name' -and $r[6].Type -eq 'String' -and $r[6].Count -eq 2)
-	assert ($r[-2].Name -eq 'Version' -and $r[-2].Type -eq 'String' -and $r[-2].Count -eq 1)
-	assert ($r[-1].Name -eq 'Version' -and $r[-1].Type -eq 'Version' -and $r[-1].Count -eq 1)
+	# another PSObject with Name and Version, both strings
+	$ps2 = 1 | Select-Object Name, Version, Extra
+	$ps2.Name = 'string'
+	$ps2.Version = $Host.Version
+
+	($r = $ps1, $ps2 | Measure-Property.ps1) | Out-String
+	equals $r[0].Name ''
+	equals $r[0].Type 'PSCustomObject'
+	equals $r[0].Count 2
+	equals $r[1].Name 'Extra'
+	equals $r[1].Type '<null>'
+	equals $r[1].Count 1
+	equals $r[2].Name 'Name'
+	equals $r[2].Type 'String'
+	equals $r[2].Count 2
+	equals $r[3].Name 'Version'
+	equals $r[3].Type 'String'
+	equals $r[3].Count 1
+	equals $r[4].Name 'Version'
+	equals $r[4].Type 'Version'
+	equals $r[4].Count 1
 }
 
 # Process dictionaries as property bags
 task DictionaryDefault {
-	$r = @{p1=1; p2=2}, @{p1=2; p3=$null} | Measure-Property
-	$r
-	assert ($r.Count -eq 4)
-	assert ($r[0].Name -eq '' -and $r[0].Type -eq 'Hashtable : IDictionary' -and $r[0].Count -eq 2)
-	assert ($r[1].Name -eq 'p1' -and $r[1].Type -eq 'Int32' -and $r[1].Count -eq 2)
-	assert ($r[2].Name -eq 'p2' -and $r[1].Type -eq 'Int32' -and $r[2].Count -eq 1)
-	assert ($r[3].Name -eq 'p3' -and $r[3].Type -eq '<null>' -and $r[3].Count -eq 1)
+	($r = @{p1=1; p2=2}, @{p1=2; p3=$null} | Measure-Property) | Out-String
+	equals $r.Count 4
+	equals $r[0].Name ''
+	equals $r[0].Type 'Hashtable : IDictionary'
+	equals $r[0].Count 2
+	equals $r[1].Name 'p1'
+	equals $r[1].Type 'Int32'
+	equals $r[1].Count 2
+	equals $r[2].Name 'p2'
+	equals $r[1].Type 'Int32'
+	equals $r[2].Count 1
+	equals $r[3].Name 'p3'
+	equals $r[3].Type '<null>'
+	equals $r[3].Count 1
 }
 
 # Process dictionaries as standard objects
 task DictionaryProperty {
-	$r = @{p1=1; p2=2}, @() | Measure-Property -DictionaryProperty
-	$r
-	assert ($r[0].Name -eq '' -and $r[0].Type -eq 'Hashtable : IDictionary' -and $r[0].Count -eq 1)
-	assert ($r[1].Name -eq '' -and $r[1].Type -eq 'Object[]' -and $r[1].Count -eq 1)
-	assert ($r[2].Name -eq 'Count' -and $r[2].Type -eq 'Int32' -and $r[2].Count -eq 2)
-	assert ($r[3].Name -eq 'IsFixedSize' -and $r[3].Type -eq 'Boolean' -and $r[3].Count -eq 2)
-	assert ($r[6].Name -eq 'Keys' -and $r[6].Type -eq 'KeyCollection : ICollection' -and $r[6].Count -eq 1)
+	($r = @{p1=1; p2=2}, @() | Measure-Property -DictionaryProperty) | Out-String
+	equals $r[0].Name ''
+	equals $r[0].Type 'Hashtable : IDictionary'
+	equals $r[0].Count 1
+	equals $r[1].Name ''
+	equals $r[1].Type 'Object[]'
+	equals $r[1].Count 1
+	equals $r[2].Name 'Count'
+	equals $r[2].Type 'Int32'
+	equals $r[2].Count 2
+	equals $r[3].Name 'IsFixedSize'
+	equals $r[3].Type 'Boolean'
+	equals $r[3].Count 2
+	equals $r[6].Name 'Keys'
+	equals $r[6].Type 'KeyCollection : ICollection'
+	equals $r[6].Count 1
 }
 
 # Test case sensitive names
@@ -54,14 +83,18 @@ task CaseSensitive {
 	$data = @{n=1}, @{N=1}
 
 	# default is ignore case
-	$r = $data | Measure-Property
-	$r
-	assert ($r.Count -eq 2 -and $r[1].Name -ceq 'n' -and $r[1].Count -eq 2)
+	($r = $data | Measure-Property) | Out-String
+	equals $r.Count 2
+	equals $r[1].Name 'n'
+	equals $r[1].Count 2
 
 	# case sensitive
-	$r = $data | Measure-Property -CaseSensitive
-	$r
-	assert ($r.Count -eq 3 -and $r[1].Name -ceq 'n' -and $r[1].Count -eq 1 -and $r[2].Name -ceq 'N' -and $r[2].Count -eq 1)
+	($r = $data | Measure-Property -CaseSensitive) | Out-String
+	equals $r.Count 3
+	equals $r[1].Name 'n'
+	equals $r[1].Count 1
+	equals $r[2].Name 'N'
+	equals $r[2].Count 1
 }
 
 task Types {
@@ -80,11 +113,11 @@ public static class TestEnumerable {
 		p3_ICollection = @{}.Keys
 		p4_IEnumerable = [TestEnumerable]::Get()
 	} | Measure-Property
-	$r
-	assert ($r[0].Type -eq 'Hashtable : IDictionary')
-	assert ($r[1].Type -eq 'Object[]')
-	assert ($r[2].Type -eq 'ArrayList : IList')
-	assert ($r[3].Type -eq 'KeyCollection : ICollection')
+	$r | Out-String
+	equals $r[0].Type 'Hashtable : IDictionary'
+	equals $r[1].Type 'Object[]'
+	equals $r[2].Type 'ArrayList : IList'
+	equals $r[3].Type 'KeyCollection : ICollection'
 	if ($PSVersionTable.PSVersion.Major -ge 3) {
 		assert ($r[4].Type -like '* : IEnumerable')
 	}